requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
black==20.8b1
click==7.1.2
Django==3.1.6
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.6.0
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
